Rename CommandRowOnly file to CommandItem
